Clear reservations from holiday absence days
Also refactor and reorganize reservation/absence queries.
