+++ service/src/main/kotlin/fi/espoo/evaka/reservations/Reservations.kt
@@ -9,11 +9,21 @@ import com.fasterxml.jackson.databind.JsonSerializer
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
 import fi.espoo.evaka.shared.ChildId
-import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.BadRequest
 import fi.espoo.evaka.shared.domain.FiniteDateRange
 import java.time.LocalDate
 import java.time.LocalTime
+
+fun convertMidnightEndTime(timeRange: TimeRange) =
+    if (timeRange.endTime == LocalTime.of(0, 0).withNano(0).withSecond(0))
+        timeRange.copy(endTime = LocalTime.of(23, 59))
+    else timeRange
+
+fun validateReservationTimeRange(timeRange: TimeRange) {
+    if (timeRange.endTime <= timeRange.startTime) {
+        throw BadRequest("Reservation start (${timeRange.startTime}) must be before end (${timeRange.endTime})")
+    }
+}
 
 data class DailyReservationRequest(
     val childId: ChildId,
@@ -52,6 +62,8 @@ data class OpenTimeRange(
     val endTime: LocalTime?,
 )
 
+private fun padWithZeros(hoursOrMinutes: Int) = hoursOrMinutes.toString().padStart(2, '0')
+private fun LocalTime.format() = "${padWithZeros(hour)}:${padWithZeros(minute)}"
 class OpenTimeRangeSerializer : JsonSerializer<OpenTimeRange>() {
     override fun serialize(value: OpenTimeRange, gen: JsonGenerator, serializers: SerializerProvider) {
         gen.writeStartObject()
@@ -59,52 +71,4 @@ class OpenTimeRangeSerializer : JsonSerializer<OpenTimeRange>() {
         gen.writeObjectField("endTime", value.endTime?.format())
         gen.writeEndObject()
     }
-}
-
-private fun LocalTime.format() = "${padWithZeros(hour)}:${padWithZeros(minute)}"
-private fun padWithZeros(hoursOrMinutes: Int) = hoursOrMinutes.toString().padStart(2, '0')
-
-fun validateReservationTimeRange(timeRange: TimeRange) {
-    if (timeRange.endTime <= timeRange.startTime) {
-        throw BadRequest("Reservation start (${timeRange.startTime}) must be before end (${timeRange.endTime})")
-    }
-}
-
-fun convertMidnightEndTime(timeRange: TimeRange) =
-    if (timeRange.endTime == LocalTime.of(0, 0).withNano(0).withSecond(0))
-        timeRange.copy(endTime = LocalTime.of(23, 59))
-    else timeRange
-
-fun Database.Transaction.clearAbsencesWithinPeriod(period: FiniteDateRange, childIds: Set<ChildId>) {
-    this.createUpdate(
-        "DELETE FROM absence WHERE child_id = ANY(:childIds) AND :period @> date"
-    )
-        .bind("childIds", childIds.toTypedArray())
-        .bind("period", period)
-        .execute()
-}
-
-fun Database.Transaction.clearOldAbsences(childDatePairs: List<Pair<ChildId, LocalDate>>) {
-    val batch = prepareBatch(
-        "DELETE FROM absence WHERE child_id = :childId AND date = :date"
-    )
-
-    childDatePairs.forEach { (childId, date) ->
-        batch.bind("childId", childId).bind("date", date).add()
-    }
-
-    batch.execute()
-}
-
-fun Database.Transaction.clearOldReservations(reservations: List<Pair<ChildId, LocalDate>>) {
-    val batch = prepareBatch("DELETE FROM attendance_reservation WHERE child_id = :childId AND date = :date")
-
-    reservations.forEach { (childId, date) ->
-        batch
-            .bind("childId", childId)
-            .bind("date", date)
-            .add()
-    }
-
-    batch.execute()
 }
